% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gelse.R
\name{gelse}
\alias{gelse}
\title{gelse}
\usage{
gelse(x, v, default = NA)
}
\arguments{
\item{x}{list of n logical vectors, logical matrix, or logical data.frame}

\item{v}{vector with n elements}
}
\value{
a vector with the same type as v
}
\description{
gelse
}

